integer types normalized

+++ src/language/exceptions/interpreterExceptions.h
@@ -38,7 +38,7 @@ class ArgumentError : public LanguageException {
 
 class ParserException : public LanguageException {
  public:
-  ParserException(const string&, const string&, unsigned) noexcept;
+  ParserException(const string&, const string&, size_t) noexcept;
   ParserException(const ParserException&) = default;
 
   ParserException& operator=(const ParserException&) = default;
